Import project

+++ scond-ngc/src/main/java/br/com/cdtec/repository/CategoriaRepository.java
@@ -1,5 +1,7 @@
 package br.com.cdtec.repository;
 
+import java.math.BigInteger;
+
 import org.springframework.context.annotation.Lazy;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
@@ -9,7 +11,7 @@ import br.com.cdtec.entity.Categoria;
 
 
 @Lazy(true)
-public interface CategoriaRepository extends GenericRepository<Categoria, Long>  {
+public interface CategoriaRepository extends GenericRepository<Categoria, BigInteger>  {
 	
 	@Query("SELECT c FROM Categoria c WHERE c.dsCategoria = :dsCategoria")
 	Categoria pesquisarPorDescricao(@Param("dsCategoria") String dsCategoria);
