+++ scond-ngc/src/main/java/br/com/cdtec/service/CategoriaService.java
@@ -1,5 +1,7 @@
 package br.com.cdtec.service;
 
+import java.math.BigInteger;
+
 import org.springframework.stereotype.Service;
 
 import br.com.cdtec.crud.service.CrudService;
@@ -8,7 +10,7 @@ import br.com.cdtec.repository.CategoriaRepository;
 
 
 @Service
-public class CategoriaService extends CrudService<Categoria, Long, CategoriaRepository> {
+public class CategoriaService extends CrudService<Categoria, BigInteger, CategoriaRepository> {
 
 	private static final long serialVersionUID = 1L;
